% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Leukemia.data}
\alias{Leukemia.data}
\title{Survival Times of 33 Leukemia Patients}
\format{A data frame with 33 observations on the following 3 variables.
 \describe{
   \item{\code{x1}}{a numeric vector (covariate)}
   \item{\code{x2}}{a numeric vector (covariate)}
   \item{\code{y}}{a numeric vector (the dependent variable)}
 }}
\source{
Feigl, P., and Zelen, M. (1965), \emph{Estimation of Exponential Survival Probabilities With Concomitant Information}, Biometrics, 21, pp. 826-838.
        Ibrahim, Joseph G. (1990). \emph{Incomplete Data in Generalized Linear Models}. Journal of the American Statistical Association, Vol.85, No. 411, pp. 765 - 769.
}
\usage{
data(Leukemia.data)
}
\description{
This data set is taken from Ibrahim, Joseph G. (1990). \emph{Incomplete Data in Generalized Linear Models}. Journal of the American Statistical Association, Vol.85, No. 411, pp. 765 - 769. The original source is Feigl, P., and Zelen, M. (1965), \emph{Estimation of Exponential Survival Probabilities With Concomitant Information}, Biometrics, 21, pp. 826-838.
}

