% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{TLI.data}
\alias{TLI.data}
\title{TLI Study of 82 Patients}
\format{A data frame with 82 observations on the following 4 variables (3 independent variables and 1 dependent variable).
 \describe{
 \item{\code{x1}}{a numeric vector}
 \item{\code{x2}}{a numeric vector}
 \item{\code{x3}}{a numeric vector}
 \item{\code{y}}{a numeric vector (the dependent variable)}
 }}
\source{
Colice, G. L., Stukel, T. A., and Dain, B. (1989), \emph{Laryngeal Complications of Prolonged Intubation}, Chest, No. 96, pp. 877-884.
        Ibrahim, Joseph G. (1990). \emph{Incomplete Data in Generalized Linear Models}. Journal of the American Statistical Association, Vol.85,
}
\usage{
data(TLI.data)
}
\description{
This data set is taken from Ibrahim, Joseph G. (1990). \emph{Incomplete Data in Generalized Linear Models}. Journal of the American Statistical Association, Vol.85, No. 411, pp. 765 - 769. The original source is Colice, G. L., Stukel, T. A., and Dain, B. (1989), \emph{Laryngeal Complications of Prolonged Intubation}, Chest, No. 96, pp. 877-884.
}

